Fix tests for CI, Move Data Classes
Done
- Fixed the tests we had for our UI for our CI so that they should pass now as they are actually testing the updated information on the pages.
- Moved the data classes as they were in the ui.theme package and that was causing issue with my ability to compile.
