+++ src/android/GetGrinnected/MyApplication/app/src/main/java/com/example/myapplication/Data.kt
@@ -1,4 +1,4 @@
-package com.example.myapplication.ui.theme
+package com.example.myapplication
 
 data class Data(
     val AllDay: Boolean,
