+++ src/android/GetGrinnected/MyApplication/app/src/main/java/com/example/myapplication/EventData.kt
@@ -1,4 +1,4 @@
-package com.example.myapplication.ui.theme
+package com.example.myapplication
 
 data class EventData(
     val `data`: List<Data>
